feat: add matter device mapper of node_id and device_id
