+++ src/panels/config/integrations/integration-panels/matter/show-dialog-matter-node-binding.ts
@@ -1,10 +1,12 @@
 import { fireEvent } from "../../../../../common/dom/fire_event";
 import type { MatterNodeBinding } from "../../../../../data/matter";
+import type { MatterDeviceMapper } from "./matter-binding-node-device-mapper";
 
 export interface MatterNodeBindingDialogParams {
   device_id: string;
   onUpdate: (bindings: Record<string, MatterNodeBinding[]>) => void;
   bindings: Record<string, MatterNodeBinding[]>;
+  deviceMapper: MatterDeviceMapper;
 }
 
 export const loadMatterNodeBindingDialog = () =>
